requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
jsonschema==4.19.0
google-generativeai==0.3.2
beautifulsoup4==4.12.3
requests==2.31.0
xlsxwriter==3.1.9 